work done on the app:     * the architecture is mostly there, some tweaks here and there and still some work is needed to make it happen.     * need to work on the unit testing, only skeleton code is commited.
work done on the report: added a chapter on the design of the application but maybe it will be merged with the other work related chapters.

work done on the presentation: sectioning is probably done with but I think I will add two or three frame, some work in the report will be added.